refactor: [#1221] move core statistics mod to statistics context
