+++ src/core/statistics/mod.rs
@@ -28,3 +28,5 @@ pub mod event;
 pub mod keeper;
 pub mod metrics;
 pub mod repository;
+pub mod services;
+pub mod setup;
