+++ src/core/statistics/setup.rs
