+++ src/servers/apis/v1/context/stats/resources.rs
@@ -2,7 +2,7 @@
 //! API context.
 use serde::{Deserialize, Serialize};
 
-use crate::core::services::statistics::TrackerMetrics;
+use crate::core::statistics::services::TrackerMetrics;
 
 /// It contains all the statistics generated by the tracker.
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
@@ -121,8 +121,8 @@ mod tests {
     use torrust_tracker_primitives::torrent_metrics::TorrentsMetrics;
 
     use super::Stats;
-    use crate::core::services::statistics::TrackerMetrics;
     use crate::core::statistics::metrics::Metrics;
+    use crate::core::statistics::services::TrackerMetrics;
 
     #[test]
     fn stats_resource_should_be_converted_from_tracker_metrics() {
